requirements.txt
#fastapi
#uvicorn
face-recognition
#python-multipart
#pillow
#celery
#redis
#psycopg2-binary
pytest
pillow
